backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
environs==11.0.0
exceptiongroup==1.2.2
fastapi==0.112.2
fastapi_cors==0.0.6
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langsmith==0.1.111
marshmallow==3.22.0
multidict==6.0.5
mypy-extensions==1.0.0
neo4j==5.24.0
numpy==1.26.4
openai==1.43.0
orjson==3.10.7
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.4
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
yarl==1.9.8
